added 404 exception

+++ db/db_article.py
@@ -1,9 +1,12 @@
+from fastapi import HTTPException, status
 from sqlalchemy.orm.session import Session
-from sqlalchemy.orm import joinedload
 from db.models import DbArticle
+from exceptions import StoryException
 from schemas import ArticleBase
 
 def create_article(request: ArticleBase, db: Session):
+    if request.content.startswith("Il etait une fois, "):
+        raise StoryException("No story here please.")
     new_article = DbArticle(
         title = request.title,
         content = request.content,
@@ -18,15 +21,25 @@ def create_article(request: ArticleBase, db: Session):
 
 def read_all(db: Session):
     article = db.query(DbArticle).all()
+    if not article:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"No article found.")
     return article
+
 
 def read_article(id: int, db: Session):
     article = db.query(DbArticle).filter(DbArticle.id == id).first()
+    if not article:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Article id {id} not found.")
     return article
 
 
 def update_article(id: int, request: ArticleBase, db: Session):
     article = db.query(DbArticle).filter(DbArticle.id == id)
+    if not article.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Article id {id} not found.")
     article.update({
         DbArticle.title: request.title,
         DbArticle.content: request.content,
@@ -40,6 +53,9 @@ def update_article(id: int, request: ArticleBase, db: Session):
 
 def delete_article(id: int, db: Session):
     article = db.query(DbArticle).filter(DbArticle.id == id).first()
+    if not article:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Article id {id} not found.")
     db.delete(article)
     db.commit()
     return {"data": article}
